Introduce new FooterView stack.
KI: When the device is locked, the "Unlock to see older
notifications" message is showing, and we receive a notification,
the clear all button is showing on top of the message. This should be
fixed in the next CL when we actually handle the visibility of the clear
all button in the new stack.

Bug: 293167744
Test: atest FooterViewTest FooterViewModelTest
NotificationStackScrollLayoutTest
also tested manually with the flag on + off, and ran scenario tests with the flag on in another CL
Flag: [type: ACONFIG] [com.android.systemui.notifications_footer_view_refactor] [state: DEVELOPMENT]

Change-Id: I6be76b51f088a77efeab4e4908f223d7d4940230