Call mergeWithNextTransaction in registerCallbacksForSync
In some cases, syncNextTransaction is called but not to sync a buffer
and only to get called back when draw is complete. In that scenario, the
pendingTransaction should be merged with the draw and not the SSG.
Change the code to always mergeWithNextTransaction since that will
handle both cases when buffer is synced and not.

Test: Wallpaper picker cropping
Fixes: 303006907
Change-Id: I752ba2ce314a24962e8c55d34b233b1e4961c6ef