Fix NPE in IMM setInteractive
This fixes a null pointer exception in InputMethodManager when handling
a setInteractive call, which was triggered when the root view of the
focused window has been removed between posting this call on the
UI thread and actually handling it.

This also improves the performance of this call by first checking if we
are already on the UI thread, in which case we can handle it directly.

Bug: 283146523
Test: n/a
Change-Id: I1a0734bbb7e5b4b28f1dbe73f7e7a3e8e4441396