uinput: move Event.Reader to its own file
We'll shortly be adding a parser for a different file format (evemu),
and having the different parsers in separate files will make the code
nicer to navigate.

In addition to moving it to a separate file, this CL renames Reader to
JsonStyleParser to avoid confusion with Java's Reader classes, and
replaces some switch statements with the new-style switch syntax to
improve readability.

Bug: 302297266
Test: atest CtsInputHostTestCases
Change-Id: Id6c7536a2caea3b61fac615dce071129bf4dc1e4