Introduce movable cutout feature flag
This flag is introduced to develop a support of movable display cutout.
With the feature, the device owners can define cutout preferred edges
for different rotations through device config file.

The preferred edge will override the other edge decision logic in the
system. It's device owner's responsibility to make sure the values are
reasonable.

Bug: 302387383
Test: m
Change-Id: I50f32ebc681a4bb051818130fc7d82fdb6293cdd