Move AlternateBouncer scrim to the KeyguardRootView
So that the alternate bouncer scrim will render
underneath the device entry icon view. This handles
the case when the alternate bouncer shows on the lockscreen
WITHOUT the shade expanded.

Also, when the device_entry_udfps_refactor flag is enabled,
the AlternateBouncer availability is driven by the supported
fingerprint sensor type on the device instead of by
the (Udfps/SideFPS) controllers.

Note: click listener stopped working when the alternate bouncer
view was moved. Alternate bouncer needs to support swiping gesture
in the future, so the touch gesture/click will be updated in a
future CL.

Test: atest AlternateBouncerInteractorTest
Test: atest SystemUITests
Bug: 287599719
Bug: 305234447
Flag: ACONFIG com.android.systemui.device_entry_udfps_refactor DEVELOPMENT
Change-Id: Ibfe3474606d381d0cdc85242c0714e006b3fb176