Apply Night Mode as part of ZenDeviceEffects
(Using a temporary API until the new one from b/313418335 is available).

Update DefaultDeviceEffectsApplier to only apply the "diff" of effects (less service calls).

This CL expands the usage of the "origin enum" to cover all the updates of the ZenModeConfig through ZenModeHelper (including setAZRState, loading, etc) and reorders arguments to make them consistent. The enum was also moved to ZenModeConfig so it can be used in DeviceEffectsApplier.

Bug: 308673343
Test: atest ZenModeHelperTest (unit)
Change-Id: I9b7599efdc8f1a55ed026b371d1be639fadf133a