Trigger touch boost when the viewVisibility of ViewRootImpl changed
This main purpose of this CL is to trigger touch boost when when the viewVisibility of ViewRootImpl is changed and the visibility is visible.

Also, having frame rate boost when a ViewRootImpl is first available.

This can ensure the transitions / animations are smooth after a laucher
is lauching a new window.

Test: atest ViewRootImplTest
Bug: 315141420
Change-Id: Ib61775ed4c13142d73bdbe6a577fe61e2635984e