Looper: Add sysprop for verbose slow message logging
By default, if a looper detects a slow delivery it will suppress all
further logs until it starts to deliver messages in time again. This
patch adds a verboseLogging flag, which can be enabled by a sysprop,
that disables this suppression, thereby making the slow delivery logging
more verbose.

Having the verbose logging makes it easier to understand and debug
issues related to slow delivery since it allows one to see exactly which
and how many messages get delivered slowly by a looper.

Bug: 326356810
Test: manual - Setprop log.looper.slow.verbose and observe log
Based-on-patch-by: Nicklas Ekstrand <nicklas.ekstrand@sonymobile.com>
Change-Id: Ibe273c27b7e11b20aa79347e529cdec1f69e59c6