[pm] fix corner cases of DELETE_KEEP_USER with multi-user
Fixing the corner case when DELETE_KEEP_USER is used in one user but not
in another user. The key is that before we decide to wipe a
PackageSetting, we need to know if the deletion was meant for all users
or a specific single user. For the later case, we need to check if the
data still exists on other users.

BUG: 314207341
Test: manual with steps listed in the bug (will add CTS)
Test: atest CtsPackageInstallTestCases
Test: atest CtsPackageManagerTest

Change-Id: I42c6727e544c63fb5618ca8d6bd46ae5d295833d