Share the Picture object of MovableElement
This CL makes sure that MovableElements share their Picture object.
Given that a MovableElement is always composed in a scene, there is no
need to have multiple Picture objects. This also ensures that the
picture contains the last drawing commands of the movable element,
independently of which scenes it is composed in.

Bug: 291053742
Test: Manual
Flag: N/A
Change-Id: I32bd35754ce701c01712612649b6c43dfdd8073f