Remove need for isComplete, but still always send final animation value
Due to not using a SharedFlow anymore, there is a possibility for
collectors of this flow to not receive the final value due to the
outer-scoped isComplete value. The possibility of this increases
if there are multiple consumers of the same animation.

Removing isComplete removes this possibility. The final value will
now eventually show up to all consumers, but it is not guaranteed
to be arrive to all of them at the same time.

DistinctUntilChanged prevents the final value from being repeatedly
emitted to the consumers.

Bug: 318543679
Flag: NONE
Test: atest KeyguardTransitionAnimationFlowTest.kt
Test: manual - observe final values coming in to both shortcuts with keyguard_bottom_area_refactor flag turned on
Change-Id: I781f90cb80fcf58bec69c258a206ab717bdd2022